Add files via upload
This current test build has a map, a player character, and collision. Its a low detail map that just has the basics so I could get collision working. And the player currently only faces downward, as I needed to focus on getting the animation to work, not all of them, just one to make it work. This test build does require Java 21, which I have installed on my personal computers.